+++ flocker/node/functional/test_deploy.py
@@ -7,11 +7,13 @@ Functional tests for ``flocker.node._deploy``.
 from subprocess import check_call
 
 from twisted.trial.unittest import TestCase
+from twisted.python.filepath import FilePath
 
-from .. import Deployer, Deployment, Application, DockerImage, Node
+from .. import (
+    Deployer, Deployment, Application, DockerImage, Node, AttachedVolume, Link)
 from ..gear import GearClient
 from ..testtools import wait_for_unit_state, if_gear_configured
-from ...testtools import random_name
+from ...testtools import random_name, DockerImageBuilder, assertContainsAll
 from ...volume.testtools import create_volume_service
 from ...route import make_memory_network
 
@@ -37,7 +39,9 @@ class DeployerTests(TestCase):
                  applications=frozenset([Application(
                      name=name,
                      image=DockerImage.from_string(
-                         u"openshift/busybox-http-app"))]))]))
+                         u"openshift/busybox-http-app"),
+                     links=frozenset(),
+                     )]))]))
 
         d = deployer.change_node_state(desired_state,
                                        Deployment(nodes=frozenset()),
@@ -61,4 +65,125 @@ class DeployerTests(TestCase):
                                                     [u'active']))
 
         # Test will timeout if unit was not restarted:
+        return d
+
+    @if_gear_configured
+    def test_environment(self):
+        """
+        The environment specified in an ``Application`` is passed to the
+        container.
+        """
+        docker_dir = FilePath(__file__).sibling('env-docker')
+        image = DockerImageBuilder(test=self, source_dir=docker_dir)
+        image_name = image.build()
+
+        application_name = random_name()
+
+        gear_client = GearClient("127.0.0.1")
+        self.addCleanup(gear_client.remove, application_name)
+
+        volume_service = create_volume_service(self)
+        deployer = Deployer(volume_service, gear_client,
+                            make_memory_network())
+
+        expected_variables = frozenset({
+            'key1': 'value1',
+            'key2': 'value2',
+        }.items())
+
+        desired_state = Deployment(nodes=frozenset([
+            Node(hostname=u"localhost",
+                 applications=frozenset([Application(
+                     name=application_name,
+                     image=DockerImage.from_string(
+                         image_name),
+                     environment=expected_variables,
+                     volume=AttachedVolume(
+                         name=application_name,
+                         mountpoint=FilePath('/data'),
+                         ),
+                     links=frozenset(),
+                     )]))]))
+
+        d = deployer.change_node_state(desired_state,
+                                       Deployment(nodes=frozenset()),
+                                       u"localhost")
+        d.addCallback(lambda _: wait_for_unit_state(gear_client,
+                                                    application_name,
+                                                    [u'active']))
+
+        def started(_):
+            volume = volume_service.get(application_name)
+            path = volume.get_filesystem().get_path()
+            contents = path.child(b'env').getContent()
+
+            assertContainsAll(
+                haystack=contents,
+                test_case=self,
+                needles=['{}={}\n'.format(k, v)
+                         for k, v in expected_variables])
+        d.addCallback(started)
+        return d
+
+    @if_gear_configured
+    def test_links(self):
+        """
+        The links specified in an ``Application`` are passed to the
+        container as environment variables.
+        """
+        docker_dir = FilePath(__file__).sibling('env-docker')
+        image = DockerImageBuilder(test=self, source_dir=docker_dir)
+        image_name = image.build()
+
+        application_name = random_name()
+
+        gear_client = GearClient("127.0.0.1")
+        self.addCleanup(gear_client.remove, application_name)
+
+        volume_service = create_volume_service(self)
+        deployer = Deployer(volume_service, gear_client,
+                            make_memory_network())
+
+        expected_variables = frozenset({
+            'ALIAS_PORT_80_TCP': 'tcp://localhost:8080',
+            'ALIAS_PORT_80_TCP_PROTO': 'tcp',
+            'ALIAS_PORT_80_TCP_ADDR': 'localhost',
+            'ALIAS_PORT_80_TCP_PORT': '8080',
+        }.items())
+
+        link = Link(alias=u"alias",
+                    local_port=80,
+                    remote_port=8080)
+
+        desired_state = Deployment(nodes=frozenset([
+            Node(hostname=u"localhost",
+                 applications=frozenset([Application(
+                     name=application_name,
+                     image=DockerImage.from_string(
+                         image_name),
+                     links=frozenset([link]),
+                     volume=AttachedVolume(
+                         name=application_name,
+                         mountpoint=FilePath('/data'),
+                         ),
+                     )]))]))
+
+        d = deployer.change_node_state(desired_state,
+                                       Deployment(nodes=frozenset()),
+                                       u"localhost")
+        d.addCallback(lambda _: wait_for_unit_state(gear_client,
+                                                    application_name,
+                                                    [u'active']))
+
+        def started(_):
+            volume = volume_service.get(application_name)
+            path = volume.get_filesystem().get_path()
+            contents = path.child(b'env').getContent()
+
+            assertContainsAll(
+                haystack=contents,
+                test_case=self,
+                needles=['{}={}\n'.format(k, v)
+                         for k, v in expected_variables])
+        d.addCallback(started)
         return d