+++ flocker/volume/_ipc.py
@@ -17,13 +17,27 @@ from characteristic import with_cmp
 
 from zope.interface import Interface, implementer
 
+from twisted.internet.defer import succeed
+
 from .service import DEFAULT_CONFIG_PATH
+from .filesystems.zfs import Snapshot
 
 
 class IRemoteVolumeManager(Interface):
     """
     A remote volume manager with which one can communicate somehow.
     """
+    def snapshots(volume):
+        """
+        Retrieve a list of the snapshots which exist for the given volume.
+
+        :param Volume volume: The volume for which to retrieve snapshots.
+
+        :return: A ``Deferred`` that fires with a ``list`` of ``Snapshot``
+            instances giving the snapshot information.  The snapshots are
+            ordered from oldest to newest.
+        """
+
     def receive(volume):
         """
         Context manager that returns a file-like object to which a volume's
@@ -63,6 +77,24 @@ class RemoteVolumeManager(object):
         self._destination = destination
         self._config_path = config_path
 
+    def snapshots(self, volume):
+        """
+        Run ``flocker-volume snapshots`` on the destination and parse the
+        output into a ``list`` of ``Snapshot`` instances.
+        """
+        data = self._destination.get_output(
+            [b"flocker-volume",
+             b"--config", self._config_path.path,
+             b"snapshots",
+             volume.uuid.encode("ascii"),
+             volume.name.encode("ascii")]
+        )
+        return succeed([
+            Snapshot(name=name)
+            for name
+            in data.splitlines()
+        ])
+
     def receive(self, volume):
         return self._destination.run([b"flocker-volume",
                                       b"--config", self._config_path.path,
@@ -90,6 +122,12 @@ class LocalVolumeManager(object):
         :param VolumeService service: The service to communicate with.
         """
         self._service = service
+
+    def snapshots(self, volume):
+        """
+        Interrogate the volume's filesystem for its snapshots.
+        """
+        return volume.get_filesystem().snapshots()
 
     @contextmanager
     def receive(self, volume):